+++ benchmark/ingestion/load_tpch.sql
@@ -1,4 +1,4 @@
-CALL dbgen(sf=1);
+CALL dbgen(sf=1::DOUBLE);
 EXPORT DATABASE 'duckdb_benchmark_data/tpch_parquet' (FORMAT PARQUET);
 EXPORT DATABASE 'duckdb_benchmark_data/tpch_csv' (FORMAT CSV);
