+++ benchmark/tpch_plan_cost/init/load.sql
@@ -1,2 +1,2 @@
-CALL dbgen(sf=1)
+CALL dbgen(sf=1::DOUBLE)
